==> weakSupervision/Snorkel/Step1_Preprocess_Data_for_Snorkel.ipynb <==
import pandas as pd
import os
import json
import numpy as np

data_path = "/efs/CONSORT2/data/parsed_data/"
files_list = os.listdir(data_path)
number_files = len(files_list)

PMIDs_list = []
for file_name in files_list:
    PMID = file_name.replace(".json","") 
    PMIDs_list.append(PMID)
print ("Number of xmls parsed: ", len(PMIDs_list))

data_file = "/efs/CONSORT2/data/snorkel_data/all_data_05182021.csv"
data_df = pd.read_csv(data_file)
print (len(data_df))

data_df_test = data_df[(data_df["PMID"] == 25515666)]

data_df_test = data_df_test[data_df_test["sid"] == 0]
with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    display(data_df_test)
test =  (data_df_test["sentence"])
print (test)

# Manual annotated PMIDs, using converter to get PMIDs
manual_data_file = "/efs/CONSORT2/data/snorkel_data/all_manual_PMIDs.csv"
manual_data_df = pd.read_csv(manual_data_file)
manual_PMIDs = set(manual_data_df["PMID"])
manual_PMIDs = list(manual_PMIDs)
manual_str_PMIDs = []
for PMID in manual_PMIDs:
    PMID = str(PMID)
    manual_str_PMIDs.append(PMID)

#exclude manual annotated PMCIDs from the list
PMIDs_not_in_manual_list = [x for x in PMIDs_list if x not in manual_str_PMIDs]
print ("Number of PMIDs after removing manual data set: ", len(PMIDs_not_in_manual_list))

PMIDs_in_manual_excluded_list = [x for x in manual_str_PMIDs if x not in PMIDs_list]

# Read RCT Outputs
RCT_Tagger_output = "/efs/MultiTagger/consort_RANDOMIZED_CONTROLLED_TRIAL.out"
RCT_Tagger_output_df = pd.read_csv(RCT_Tagger_output, sep = "\t")
print ("Number of papers output from Tagger: ", len(RCT_Tagger_output_df))

# Filter the ones that we actually parse XMLs
RCT_Tagger_output_df = RCT_Tagger_output_df[RCT_Tagger_output_df["PMID"].isin(PMIDs_list)]
print ("Number of papers output from Tagger that we parsed: ", len(RCT_Tagger_output_df))

display(RCT_Tagger_output_df.head(100))

RCT_predicted_df = RCT_Tagger_output_df[(RCT_Tagger_output_df["PREDICTION"] =="RANDOMIZED_CONTROLLED_TRIAL")]
print (len(RCT_predicted_df))

display(RCT_predicted_df)

import random 

RCT_predicted_df = RCT_Tagger_output_df[(RCT_Tagger_output_df["PREDICTION"] =="RANDOMIZED_CONTROLLED_TRIAL") & 
                                        (RCT_Tagger_output_df["CONFIDENCE"] >=0.95)]

PMID_RCT_predicted = set(RCT_predicted_df["PMID"])
PMID_RCT_predicted = list(PMID_RCT_predicted)

PMID_RCT_str_predicted = []
for PMID in PMID_RCT_predicted:
    PMID = str(PMID)
    PMID_RCT_str_predicted.append(PMID)
    
print ("Number of RCTs from Tagger we are sampling: ", len(PMID_RCT_str_predicted))

# Check titles
import pandas as pd
import os
import random
import collections
import urllib
from Bio import Entrez
from xml.dom import minidom
import xml.etree.ElementTree as ET


bad_PMIDs = []

PMID_list = []
Title_list = []
Publication_Types_list = []
count = 0

for PMID in PMID_RCT_str_predicted:
    if PMID != "":
        article_pmid = PMID
        article_title = ""
        pt_lst = []
        
        if count%500 == 0:
            print (count)

        link = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&api_key=1a7d18b4a3b452b6f1d688fa1ad4282bfa09&retmode=xml&id=" + str(PMID)

        request = urllib.request.Request(link) 
            
        try:
            result = urllib.request.urlopen(request)
        except:
            bad_PMIDs.append(PMID)
            pass
            
        try:
            tree = ET.parse(result)
        except ET.ParseError as e:
            pass
                        
        doc = tree.getroot()

        #get title
        title = doc.find (".//ArticleTitle")
        if title is None:
            title_sub = doc.find (".//ArticleTitle/i")
            if title_sub is None:
                article_title = "None"
            else:
                article_title = title_sub.text
        else:
            article_title = title.text
            
              
        # get publication type
        publication_type_list = doc.find(".//PublicationTypeList")
        if publication_type_list is not None:      
            publicationtypes = publication_type_list.getchildren()
            for publication_type in publicationtypes:
                pt_lst.append(publication_type.text)
        else:
            pt_lst.append("None")
            
        PMID_list.append(article_pmid)
        Title_list.append(article_title)
        Publication_Types_list.append(pt_lst)
        
        count+=1 
        
all_data_withmetadata_df = pd.DataFrame(list(zip(PMID_list,Title_list, Publication_Types_list)), 
                   columns =['PMID','Title','Publication Types']) 


print ("Number of PMIDs that don't have links: ", len(bad_PMIDs))
print ("Number of PMIDs that have links: ", len(all_data_withmetadata_df))

display(all_data_withmetadata_df.head(10))

# Get papers from the list
all_data_df = pd.DataFrame()
count = 0
for PMCID in PMID_RCT_str_predicted:
    if count%500 == 0:
        print (count)
    
    PMCID_df = pd.DataFrame()
    with open(data_path + PMCID + ".json") as f:
        data = json.load(f)  
        for sentence_item in data:            
            sentence_df = pd.DataFrame.from_dict(sentence_item,orient='index')
            sentence_df= sentence_df.transpose()

            PMCID_df = pd.concat([PMCID_df,sentence_df])
            
    PMCID_df["PMID"] = PMCID
    PMCID_df = PMCID_df[["PMID","sentence","sid","span","section","token","lemma","pos"]]
    PMCID_df = PMCID_df.reset_index(drop=True)

    all_data_df = pd.concat([all_data_df,PMCID_df])
    all_data_df = all_data_df.reset_index(drop=True)
    count +=1 
display(all_data_df.head(100)) 

all_data_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_data_05182021.csv", index = False)

all_data_df = pd.read_csv("/efs/CONSORT2/data/snorkel_data/all_data_05182021.csv")
display(all_data_df.head(10))
print (len(all_data_df))
PMIDs = set(all_data_df["PMID"])
print (len(PMIDs))

titles_df = all_data_df[all_data_df["sid"] == 0]
propocols = ["propocol", "Protocol"]
protocol_titles_df = titles_df[titles_df["sentence"].str.contains("protocol|Protocol")]
display(protocol_titles_df)
protocol_PMIDs = set (protocol_titles_df["PMID"])
print (len(protocol_PMIDs))

none_protocol_data_df = all_data_df[~all_data_df["PMID"].isin(protocol_PMIDs)]
print (len(none_protocol_data_df))

method_sentences_df = none_protocol_data_df[all_data_df["section"].str.contains("methods|Methods|method|Method")]
method_sentences_df_PMIDs = set(method_sentences_df["PMID"])
print (len(method_sentences_df_PMIDs))
display(method_sentences_df)

method_sentences_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_methodological_data_05212021.csv", index = False)

snorkel_data_file = "/efs/CONSORT2/data/snorkel_data/all_methodological_data_05212021.csv"

snorkel_data_df = pd.read_csv(snorkel_data_file)
snorkel_data_df = snorkel_data_df[["PMID","sid", "sentence","section"]]
snorkel_data_df.columns = ["PMID","sid", "text","section_hierarchy"]
display(snorkel_data_df.head(5))

print (len(snorkel_data_df))

# Preprocessing snorkel data

# Remove rows that are not string
snorkel_data_df = snorkel_data_df[~snorkel_data_df["text"].isnull()]
print (len(snorkel_data_df))

# Remove rows with too short text
snorkel_data_df = snorkel_data_df[~snorkel_data_df['text'].apply(lambda x: len(x.split(' ')) < 5)]
print (len(snorkel_data_df))

# Remove rows that only numbers
snorkel_data_df = snorkel_data_df[~snorkel_data_df['text'].str.isdigit()]
print (len(snorkel_data_df))

snorkel_data_df = snorkel_data_df.reset_index()
snorkel_data_df = snorkel_data_df[["PMID","sid","text","section_hierarchy"]]

# Preprocess convert all to lower case
snorkel_data_df["text"] = snorkel_data_df["text"].str.lower()
snorkel_data_df["section_hierarchy"] = snorkel_data_df["section_hierarchy"].str.lower()

display(snorkel_data_df.head(10))

# Remove sentences that are section header
preprocessed_snorkel_data_df = snorkel_data_df[~(snorkel_data_df["text"].isin(snorkel_data_df["section_hierarchy"]))]

# Remove sentences that are Figure/Table
preprocessed_snorkel_data_df = preprocessed_snorkel_data_df[~(preprocessed_snorkel_data_df["text"].str.contains("figure|table|shown in"))]
print (len(preprocessed_snorkel_data_df))

display(preprocessed_snorkel_data_df.head(100))

preprocessed_snorkel_data_df = preprocessed_snorkel_data_df.reset_index()
preprocessed_snorkel_data_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_methodological_data_preprocessed_05232021.csv", index = False)
