==> weakSupervision/Snorkel/Step2_Get_Similarity_Scores.ipynb <==
import pandas as pd
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize 
import numpy as np
import re
from numpy import dot
from numpy.linalg import norm
import math
import random

stop_words = set(stopwords.words('english')) 

# Extracting word vectors from biowordvec

biword2vec_text_file = "/efs/CONSORT2/BioWordVec/BioWordVec_PubMed_MIMICIII_d200.txt"

embeddings_dict = {}

N= 16545452

with open(biword2vec_text_file,buffering=100000) as infile:   
    next(infile)
    for i in range(N):
        line = next(infile).strip()
        items = line.split()
        word = items[0]
        vector = np.asarray(items[1:], "float32")
        embeddings_dict[word] = vector

# Functions to get sentence embeddings
def avg_feature_vector(sentence,embedding_dict): 
    sentence_embedding = np.zeros((200, ), dtype='float32')
    
    words = preprocess(sentence)
    
    for word in words:
        if word in embedding_dict.keys():
            word_embedding = embedding_dict[word]
            sentence_embedding = np.vstack((sentence_embedding,word_embedding))

    avg_vector = np.mean(sentence_embedding, axis=0)
    return avg_vector

def preprocess(string):
    string = str(string)
    string = re.sub(r'[^\w\s]','',string)
    if string != "":
        string = string.lower()
        tokenized = word_tokenize(string)
        tokens_without_sw = [word for word in tokenized if not word in stop_words]
    else:
        tokens_without_sw = []
    return tokens_without_sw


# Functions to get consine similarities and top 5

def get_cosine(vec1, vec2):
    return dot(vec1, vec2)/(norm(vec1)*norm(vec2))

import operator
def get_top_n_similarities(N,a):
    # Indices of N largest elements in list 
    # using sorted() + lambda + list slicing 
    res = sorted(range(len(a)), key=lambda i: a[i])[-N:]
    scores = []
    for index in res:
        scores.append(a[index])
    return res,scores

core_sentences = "/efs/CONSORT/skr-consort/datasets/TrainingDatasets/95.txt"
core_sentences_label_0 = "/efs/CONSORT/HeuristicsBasedAnnotations/label0_200.txt"

core_sentences_lst = []
core_labels_lst = []
f = open(core_sentences, "r")
for line in f.readlines():
    sentence_items = line.split("|")
    CONSORT_Item = sentence_items[0]
    sentence_text = sentence_items[2]
    sentence_text = sentence_text.replace("\n","")
    sentence_text = sentence_text.lower()
    
    core_sentences_lst.append(sentence_text)
    core_labels_lst.append(CONSORT_Item)
    
f_0 = open(core_sentences_label_0,"r")
for line in f_0.readlines():
    sentence_items = line.split("|")
    CONSORT_Item = "0"
    sentence_text = sentence_items[3]
    sentence_text = sentence_text.replace("\n","")
    sentence_text = sentence_text.lower()
    
    core_sentences_lst.append(sentence_text)
    core_labels_lst.append(CONSORT_Item)
    
    
core_sentences_df = pd.DataFrame(
    {'CONSORT_Item': core_labels_lst,
     'text': core_sentences_lst
    })

core_sentences_df['bio_word2vec'] = ""
for index, row in core_sentences_df.iterrows():
    core_sentence_text = row["text"]
    core_sentence_avg_feature_vector = avg_feature_vector(core_sentence_text, embeddings_dict)
    core_sentences_df.at[index, "bio_word2vec"] = core_sentence_avg_feature_vector
    
display(core_sentences_df)

# get sentence embeddings for snorkel data
data_file = "/efs/CONSORT2/data/snorkel_data/all_data_SNORKEL_RESULTS_keywords_sectionheaders_threshold08.csv"
all_data_df = pd.read_csv(data_file)
print (len(all_data_df))

all_data_df['bio_word2vec'] = ""
all_data_df['core_sentence_index'] = ""
all_data_df['core_sentence_similarity_score'] = ""

count = 0
for index_snorkel, row_snorkel in all_data_df.iterrows():
    if (count%500 == 0):
        print (count)
    sentence_text = row_snorkel["text"]
    sentence_avg_feature_vector = avg_feature_vector(sentence_text, embeddings_dict)
    all_data_df.at[index_snorkel, "bio_word2vec"] = sentence_avg_feature_vector
    
    similarity_scores_list = []
    
    for index, row in core_sentences_df.iterrows():
        core_sentence_avg_feature_vector = row["bio_word2vec"]
        similarity_score = get_cosine(sentence_avg_feature_vector,core_sentence_avg_feature_vector)
        if isinstance(np.float32(similarity_score), np.floating) :
            similarity_scores_list.append(similarity_score)
        else:
            similarity_scores_list.append(0)
            
    top_5_similar_indexes, top_5_similar_scores = get_top_n_similarities(95,similarity_scores_list)
    
    all_data_df.at[index_snorkel,'core_sentence_index'] = top_5_similar_indexes
    all_data_df.at[index_snorkel,'core_sentence_similarity_score'] = top_5_similar_scores
        
    count += 1
all_data_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_data_with_similar_sentences_indentified.csv")

data_file = "/efs/CONSORT2/data/snorkel_data/all_data_with_similar_sentences_indentified.csv"
data_df = pd.read_csv(data_file)
print (len(data_df))

display(data_df.head(100))

display(all_data_df.head(100))
