==> weakSupervision/Snorkel/Step3_Obtain_Sentence_Similarity_Info.ipynb <==
import pandas as pd
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize 
import numpy as np
import re
from numpy import dot
from numpy.linalg import norm
import math
import random

stop_words = set(stopwords.words('english')) 

data_file = "/efs/CONSORT2/data/snorkel_data/all_data_with_similar_sentences_BioBERT_embeddings_05242021.csv"
similarities_df = pd.read_csv(data_file, engine="python",encoding='utf-8', error_bad_lines=False)

display(similarities_df.head(5))
print(len(similarities_df))

core_sentences = "/efs/CONSORT/skr-consort/datasets/TrainingDatasets/95.txt"
core_sentences_label_0 = "/efs/CONSORT/HeuristicsBasedAnnotations/label0_200.txt"

core_sentences_lst = []
core_labels_lst = []
f = open(core_sentences, "r")
for line in f.readlines():
    sentence_items = line.split("|")
    CONSORT_Item = sentence_items[0]
    sentence_text = sentence_items[2]
    sentence_text = sentence_text.replace("\n","")
    sentence_text = sentence_text.lower()
    
    core_sentences_lst.append(sentence_text)
    core_labels_lst.append(CONSORT_Item)
    
f_0 = open(core_sentences_label_0,"r")
for line in f_0.readlines():
    sentence_items = line.split("|")
    CONSORT_Item = "0"
    sentence_text = sentence_items[3]
    sentence_text = sentence_text.replace("\n","")
    sentence_text = sentence_text.lower()
    
    core_sentences_lst.append(sentence_text)
    core_labels_lst.append(CONSORT_Item)
    
    
core_sentences_df = pd.DataFrame(
    {'CONSORT_Item': core_labels_lst,
     'text': core_sentences_lst
    })

display(core_sentences_df)

unique_labels = set(core_labels_lst)
print (unique_labels)

from collections import OrderedDict
from collections import Counter

def str_to_float_list(item):
    item = item.replace("[","")
    item = item.replace("]","")
    item = item.split(", ")
    item = [float(i) for i in item]
    return item

def str_to_int_list(item):
    item = item.replace("[","")
    item = item.replace("]","")
    item = item.split(", ")
    item = [int(i) for i in item]
    return item

count = 0
for index, row in similarities_df.iterrows():
    if count%500 == 0:
        print (count)
    row_similarity_score = row["core_sentence_similarity_score"]
    row_similarity_score = str_to_float_list(row_similarity_score)

    row_similarity_index = row["core_sentence_index"]
    row_similarity_index = str_to_int_list(row_similarity_index)

    #get top 10 similar core sentences
    top_n_similarity_indexes = row_similarity_index[85:]
    top_n_similar_core_sentences_df = core_sentences_df[core_sentences_df.index.isin(top_n_similarity_indexes)]  
    top_n_similar_core_sentences_df = top_n_similar_core_sentences_df.reindex(top_n_similarity_indexes)
    top_n_CONSORT_Items = top_n_similar_core_sentences_df["CONSORT_Item"].to_list()

    top_n_CONSORT_Items_by_frequency = Counter(top_n_CONSORT_Items)
    top_n_CONSORT_Items_by_frequency = OrderedDict(sorted(top_n_CONSORT_Items_by_frequency.items(), key = lambda kv : kv[1], reverse=True))

    highest_frequency_label = top_n_CONSORT_Items_by_frequency.popitem(last=False)[0]
    highest_similarity_label = top_n_CONSORT_Items[9]

    
    similarities_df.at[index, "highest_frequency_label_in_top10"] = highest_frequency_label
    similarities_df.at[index, "highest_similarity_label"] = highest_similarity_label
    
    count +=1
    
display(similarities_df.head(10))

similarities_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_data_with_Sentences_Similarity_Info_05262021.csv")


# Merge with the actual data for snorkel
snorkel_data_file = "/efs/CONSORT2/data/snorkel_data/all_methodological_data_preprocessed_05232021.csv"

snorkel_data_df = pd.read_csv(snorkel_data_file)
display(snorkel_data_df.head(5))
print (len(snorkel_data_df))

merged_snorkel_data_df = snorkel_data_df.merge(similarities_df, on = ["PMID", "sid"], how = "left")
print (len(merged_snorkel_data_df))

merged_snorkel_data_df = merged_snorkel_data_df[["PMID", "sid","text","section_hierarchy","highest_frequency_label_in_top10", "highest_similarity_label"]]
merged_snorkel_data_df = merged_snorkel_data_df.reset_index()
display(merged_snorkel_data_df.head(5))
merged_snorkel_data_df.to_csv("/efs/CONSORT2/data/snorkel_data/all_methodological_data_preprocessed_withSentenceSimilarityInfo_05262021.csv", index = False)
